Route of video edit was added.
