--- app/api/сutvideo/route.ts
@@ -1,44 +1,0 @@
-import { NextResponse } from 'next/server';
-import ffmpeg from 'fluent-ffmpeg';
-import fs from 'fs/promises';
-
-const FFMPEG_PATH = '/usr/bin/ffmpeg'; // путь к ffmpeg в WSL
-ffmpeg.setFfmpegPath(FFMPEG_PATH);
-
-export async function POST() {
-    try {
-        const videoPath = '/mnt/d/development/crononsoft/public/uploads/92c60d62-922c-485f-828e-5e04568a1b54.mp4';
-        const outputDir = '/mnt/d/development/crononsoft/public/uploads/frames';
-
-        await fs.mkdir(outputDir, { recursive: true });
-
-        const prefix = Date.now();
-
-        return new Promise((resolve, reject) => {
-            ffmpeg(videoPath)
-                .on('filenames', (filenames) => {
-                    console.log('Generated thumbnails:', filenames);
-                })
-                .on('end', () => {
-                    const thumbnails = [];
-                    for (let i = 1; i <= 10; i++) {
-                        thumbnails.push(`/uploads/frames/${prefix}-${i}.jpg`);
-                    }
-                    resolve(NextResponse.json({ thumbnails }));
-                })
-                .on('error', (err) => {
-                    console.error('FFmpeg error:', err);
-                    reject(NextResponse.json({ error: 'FFmpeg failed' }, { status: 500 }));
-                })
-                .screenshots({
-                    count: 10,
-                    folder: outputDir,
-                    filename: `${prefix}-%i.jpg`,
-                    size: '160x90',
-                });
-        });
-    } catch (e) {
-        console.error('Unexpected error:', e);
-        return NextResponse.json({ error: 'Internal error' }, { status: 500 });
-    }
-}
